moved users tests into separate files for user, users, user_session api requests
