+++ app/resources/v1/__tests__/users.test.js
@@ -13,14 +13,6 @@ jest.mock('../../../../lib/logger', () => function () {
 })
 
 // Fake user info to test the API
-const user = {
-  twitter: {
-    screenName: 'oluwaseun',
-    userId: '4232',
-    oauthAccessTokenKey: '461-Hvsiosdfsoafsoafedfd',
-    oauthAccessTokenSecret: 'WoZytO4kMLuafdafjdafja'
-  }
-}
 const emailUser = {
   auth0: {
     nickname: 'omoyajowo2015',
@@ -42,12 +34,9 @@ function setupMockServer () {
   app.use(express.json())
 
   app.post('/api/v1/users', users.post)
-  app.get('/api/v1/users/:user_id', users.get)
   // Set loginToken before running remaining endpoint
   app.use(setLoginToken)
-
-  app.put('/api/v1/users/:user_id', users.put)
-  app.delete('/api/v1/users/:user_id', users.delete)
+  app.get('/api/v1/users', users.get)
   return app
 }
 
@@ -71,47 +60,6 @@ describe('POST api/v1/users', function () {
       .send('')
       .then((response) => {
         expect(response.statusCode).toEqual(400)
-      })
-  })
-})
-
-describe('PUT api/v1/users/:user_id', function () {
-  const app = setupMockServer()
-
-  it('should respond with 204 No content when user credentails are updated', function () {
-    return request(app)
-      .put(`/api/v1/users/${user.twitter.screenName}`)
-      .type('json')
-      .send(JSON.stringify(user))
-      .then((response) => {
-        expect(response.statusCode).toEqual(204)
-      })
-  })
-
-  it.skip('should respond with 401 if a user attempts to put a request to a user that is not them', () => {})
-  it.skip('should respond with 204 if an admin user puts a request to another user', () => {})
-})
-
-describe('GET api/v1/users/:user_id', function () {
-  const app = setupMockServer()
-
-  it('should respond with 200 Ok when user is found', function () {
-    return request(app)
-      .get(`/api/v1/users/${user.twitter.screenName}`)
-      .then((response) => {
-        expect(response.statusCode).toEqual(200)
-      })
-  })
-})
-
-describe('DELETE api/v1/users/:user_id', function () {
-  const app = setupMockServer()
-
-  it('should respond with 204 No content when user signs out', function () {
-    return request(app)
-      .delete(`/api/v1/users/${user.twitter.screenName}`)
-      .then((response) => {
-        expect(response.statusCode).toEqual(204)
       })
   })
 })
